pyproject.toml
[build-system]

[project]
name = "imagep"
version = "0.0.0"
dependencies = [
    "PySide6",
    "pillow",
]
requires-python = ">=3.11"
authors = [
    {name = "Kyle Hershberger", email = "kyle.hershberger@gmail.com"},
]
maintainers = [
    {name = "Kyle Hershberger", email = "kyle.hershberger@gmail.com"},
]
description = "Lovely Spam! Wonderful Spam!"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE.*"]


[project.urls]
Repository = "https://github.com/khershberger/imagep"

[project.scripts]
imagep = "spam:main"

[project.gui-scripts]
imagep-gui = "spam:main"

[project.entry-points."spam.magical"]
tomatoes = "spam:main_tomatoes"